Move ActivateAnalyticsCTA and PreviewGraph and fix imports.
